+++ miamCore/src/androidMain/kotlin/com/miam/kmmMiamCore/Platform.kt
@@ -1,10 +1,10 @@
 package com.miam.kmmMiamCore
 
-import android.app.Activity
+import android.content.Context
 
 /**
  * This file set general platform (Android/Ios) settings
  * Android implementation
  */
 
-actual typealias KMMContext = Activity
+actual typealias KMMContext = Context