+++ changelog.md
@@ -1,5 +1,6 @@
 # 3.7.0
 - [FEA][AND] Add price template 
+- [Fix][AND] Use android app context
 
 # 3.6.0
 - [FIX][AND] Recipes page no long reloading after closing a dialog